refactor: add useCanvas
