+++ src/components/Drawing/hooks/useCanvas/constants.ts
